remove overrides
