+++ api/v1alpha1/workspace_types_test.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/gkampitakis/go-snaps/snaps"
 	corev1 "k8s.io/api/core/v1"
-	netv1 "k8s.io/api/networking/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/utils/pointer"
@@ -147,109 +146,9 @@ func TestNetworkRule_ServicePort(t *testing.T) {
 	}
 }
 
-func TestNetworkRule_IngressRule(t *testing.T) {
-	type args struct {
-		backendSvcName string
-	}
-	tests := []struct {
-		name    string
-		netRule *NetworkRule
-		args    args
-	}{
-		{
-			name: "✅ OK",
-			netRule: &NetworkRule{
-				Name:             "name",
-				PortNumber:       1111,
-				HTTPPath:         "/path",
-				TargetPortNumber: pointer.Int32(2222),
-				Host:             pointer.String("host"),
-				Group:            pointer.String("group"),
-				Public:           true,
-			},
-			args: args{
-				backendSvcName: "svcname",
-			},
-		},
-		{
-			name: "✅ host is nil",
-			netRule: &NetworkRule{
-				Name:             "name",
-				PortNumber:       1111,
-				HTTPPath:         "/path",
-				TargetPortNumber: pointer.Int32(2222),
-				Host:             nil,
-				Group:            pointer.String("group"),
-				Public:           true,
-			},
-			args: args{
-				backendSvcName: "svcname",
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := tt.netRule.IngressRule(tt.args.backendSvcName)
-			snaps.MatchJSON(t, got)
-		})
-	}
-}
-
-func TestNetworkRule_TraefikRoute(t *testing.T) {
-	type args struct {
-		backendSvcName       string
-		headerMiddlewareName string
-	}
-	tests := []struct {
-		name    string
-		netRule *NetworkRule
-		args    args
-	}{
-		{
-			name: "✅ public",
-			netRule: &NetworkRule{
-				Name:             "name",
-				PortNumber:       1111,
-				HTTPPath:         "/path",
-				TargetPortNumber: pointer.Int32(2222),
-				Host:             pointer.String("host"),
-				Group:            pointer.String("group"),
-				Public:           true,
-			},
-			args: args{
-				backendSvcName: "svcname",
-			},
-		},
-		{
-			name: "✅ not public",
-			netRule: &NetworkRule{
-				Name:             "name",
-				PortNumber:       1111,
-				HTTPPath:         "/path",
-				TargetPortNumber: pointer.Int32(2222),
-				Host:             pointer.String("host"),
-				Group:            pointer.String("group"),
-				Public:           false,
-			},
-			args: args{
-				backendSvcName:       "svcname",
-				headerMiddlewareName: "headers",
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := tt.netRule.TraefikRoute(tt.args.backendSvcName, tt.args.headerMiddlewareName)
-			snaps.MatchJSON(t, got)
-		})
-	}
-}
-
-func TestNetworkRulesByServiceAndIngress(t *testing.T) {
-	pathTypePrefix := netv1.PathTypePrefix
+func TestNetworkRulesByService(t *testing.T) {
 	type args struct {
 		svc corev1.Service
-		ing netv1.Ingress
 	}
 	tests := []struct {
 		name string
@@ -279,72 +178,12 @@ func TestNetworkRulesByServiceAndIngress(t *testing.T) {
 						},
 					},
 				},
-				ing: netv1.Ingress{
-					Spec: netv1.IngressSpec{
-						Rules: []netv1.IngressRule{
-							{
-								Host: "host.example.com",
-								IngressRuleValue: netv1.IngressRuleValue{
-									HTTP: &netv1.HTTPIngressRuleValue{
-										Paths: []netv1.HTTPIngressPath{
-											{
-												Path:     "/",
-												PathType: &pathTypePrefix,
-												Backend: netv1.IngressBackend{
-													Service: &netv1.IngressServiceBackend{
-														Name: "test-svc",
-														Port: netv1.ServiceBackendPort{
-															Name: "main",
-														},
-													},
-												},
-											},
-											{
-												Path:     "/aaa",
-												PathType: &pathTypePrefix,
-												Backend: netv1.IngressBackend{
-													Service: &netv1.IngressServiceBackend{
-														Name: "test-svc",
-														Port: netv1.ServiceBackendPort{
-															Number: 7778,
-														},
-													},
-												},
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "✅ no ingress",
-			args: args{
-				svc: corev1.Service{
-					ObjectMeta: v1.ObjectMeta{
-						Name: "test-svc",
-					},
-					Spec: corev1.ServiceSpec{
-						Ports: []corev1.ServicePort{
-							{
-								Name:       "main",
-								Port:       int32(7777),
-								Protocol:   corev1.ProtocolTCP,
-								TargetPort: intstr.FromInt(32000),
-							},
-						},
-					},
-				},
-				ing: netv1.Ingress{},
 			},
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got := NetworkRulesByServiceAndIngress(tt.args.svc, tt.args.ing)
+			got := NetworkRulesByService(tt.args.svc)
 			snaps.MatchSnapshot(t, got)
 		})
 	}
