+++ config/crd/bases/cosmo-workspace.github.io_workspaces.yaml
@@ -97,8 +97,6 @@ spec:
                 properties:
                   deploymentName:
                     type: string
-                  ingressName:
-                    type: string
                   mainServicePortName:
                     type: string
                   serviceName:
