+++ hack/local-run-test/templates/dev-code-server/cosmo-template.yaml
@@ -4,7 +4,6 @@ kind: Template
 metadata:
   annotations:
     workspace.cosmo-workspace.github.io/deployment: workspace
-    workspace.cosmo-workspace.github.io/ingress: ""
     workspace.cosmo-workspace.github.io/service: workspace
     workspace.cosmo-workspace.github.io/service-main-port: main
     workspace.cosmo-workspace.github.io/urlbase: ""
