+++ hack/local-run-test/templates/dev-code-server/team-a-template.yaml
@@ -5,7 +5,6 @@ metadata:
   annotations:
     cosmo-workspace.github.io/userroles: team-a-*
     workspace.cosmo-workspace.github.io/deployment: workspace
-    workspace.cosmo-workspace.github.io/ingress: ""
     workspace.cosmo-workspace.github.io/service: workspace
     workspace.cosmo-workspace.github.io/service-main-port: main
     workspace.cosmo-workspace.github.io/urlbase: ""
