+++ internal/cmd/template_test.go
@@ -97,7 +97,7 @@ var _ = Describe("cosmoctl [template]", func() {
 
 		DescribeTable("✅ success in normal context:",
 			run_test,
-			Entry(desc, "template", "generate", "--workspace", "--workspace-main-service-port-name", "main", "--serviceaccount", "hoge", "--required-vars", "HOGE:HOGEHOGE,FUGA:FUGAFUGA"),
+			Entry(desc, "template", "generate", "--workspace", "--workspace-main-service-port-name", "main", "--required-vars", "HOGE:HOGEHOGE,FUGA:FUGAFUGA"),
 			Entry(desc, "template", "generate", "--workspace", "--workspace-main-service-port-name", "main", "-o", "/tmp/test-cosmo-template"),
 			Entry(desc, "template", "generate", "--user-addon", "--set-default-user-addon", "--disable-nameprefix"),
 			Entry(desc, "template", "generate", "--user-addon", "--set-default-user-addon", "--cluster-scope", "--disable-nameprefix"),
